Split the event list page up a bit
